tests/wpt/web-platform-tests/css/css-scroll-snap/scroll-margin-visibility-check.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#scroll-margin">
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#scroll-padding">
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1708303">
<link rel="author" href="mailto:emilio@crisal.io" title="Emilio Cobos Ãlvarez">
<link rel="author" href="https://mozilla.org" title="Mozilla">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
body { margin: 0 }
#scroller {
  height: 500px;
  width: 500px;
  scrollbar-width: none;
  overflow: scroll;
}
#target {
  width: 100px;
  height: 100px;
  background-color: blue;
  scroll-margin: 100px;
  margin-left: 510px;
}
</style>

<div id="scroller">
  <div style="width: 450px; height: 450px;"></div>
  <div tabindex="0" id="target"></div>
  <div style="width: 2000px; height: 2000px;"></div>
</div>

<script>
let scroller = document.getElementById("scroller");
let target = document.getElementById("target");
promise_test(async function() {
  scroller.scrollTo(0, 0);
  await new Promise(resolve => {
      scroller.addEventListener("scroll", () => step_timeout(resolve, 0));
      document.getElementById("target").focus();
  });
  assert_true(scroller.scrollTop > 0, "Visibility check should not account for margin");
  assert_true(scroller.scrollLeft > 0, "Visibility check should not account for margin");
});
</script>
